requirements.txt
anyio==3.3.4
fastapi==0.70.0
fastapi-jwt-auth==0.5.0
idna==3.3
pydantic==1.8.2
PyJWT==1.7.1
sniffio==1.2.0
starlette==0.16.0
typing-extensions==3.10.0.2
